+++ src/main/java/seedu/address/storage/JsonAdaptedGradedComponent.java
@@ -61,19 +61,19 @@ public class JsonAdaptedGradedComponent {
             throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT,
                 MaxMarks.class.getSimpleName()));
         }
-        if (!MaxMarks.isValidMarks(Double.parseDouble(gcMaxMarks))) {
+        if (!MaxMarks.isValidMarks(Float.parseFloat(gcMaxMarks))) {
             throw new IllegalValueException(MaxMarks.MESSAGE_CONSTRAINTS);
         }
-        final MaxMarks modelMaxMarks = new MaxMarks(Double.parseDouble(gcMaxMarks));
+        final MaxMarks modelMaxMarks = new MaxMarks(Float.parseFloat(gcMaxMarks));
 
         if (gcWeightage == null) {
             throw new IllegalValueException(String.format(MISSING_FIELD_MESSAGE_FORMAT,
                 Weightage.class.getSimpleName()));
         }
-        if (!Weightage.isValidWeightage(Double.parseDouble(gcWeightage))) {
+        if (!Weightage.isValidWeightage(Float.parseFloat(gcWeightage))) {
             throw new IllegalValueException(Weightage.MESSAGE_CONSTRAINTS);
         }
-        final Weightage modelWeightage = new Weightage(Double.parseDouble(gcWeightage));
+        final Weightage modelWeightage = new Weightage(Float.parseFloat(gcWeightage));
 
         return new GradedComponent(modelGcName, modelMaxMarks, modelWeightage);
